Minor cleanup
git-svn-id: http://qeforge.qe-forge.org/svn/q-e/trunk/espresso@219 c92efa57-630b-4861-b058-cf58834340f0

+++ PW/pwcom.f90
@@ -55,7 +55,7 @@ module dynam
   !
   !    Variables needed for the dynamics
   !
-  real(kind=8) ::                                                    &
+  real(kind=DP) ::                                                   &
        amass(ntypx),   &! mass of atoms
        dt,             &! time step
        temperature,    &! starting temperature
